refactor: remove common types to separate file
