+++ react-components/src/types.ts
@@ -4,3 +4,14 @@ export interface IOrder {
   office: string;
   pack: boolean;
 }
+
+export interface IBook{
+  name: string;
+  author: string;
+  img: string;
+  likes: number;
+  comments: number;
+  price: number;
+}
+
+export type DetailType = 'price' | 'comments' | 'likes';
